+++ app/(tabs)/ScheduleTab.tsx
@@ -3,7 +3,6 @@ import { styles } from "./styles";
 import { AntDesign } from "@expo/vector-icons";
 import { useColorScheme } from "react-native";
 import Colors from "@/constants/Colors";
-import SwipeableButton from "./Test";
 
 export default function ScheduleScreen() {
   const colorScheme = useColorScheme();
@@ -18,7 +17,6 @@ export default function ScheduleScreen() {
           color={Colors[colorScheme ?? "light"].accent}
         />
       </ThemedButton>
-      <SwipeableButton />
     </ThemedView>
   );
 }
