--- app/createRoutine/index.tsx
@@ -1,1 +1,0 @@
-export default () => null;
